fix(agent): restore claude payload option fallback and split conformance scenarios
- Restore the payload optionId fallback for non-approval Claude interactive
  replies so a choice carried only in the payload is still forwarded.
- Move interactive response scenarios into conformance/interactive_response.go
  to keep conformance.go under the file-length baseline.
