+++ packages/agent/daemon/runtime/claude_sdk_interactive.go
@@ -35,6 +35,9 @@ func (a *ClaudeCodeSDKAdapter) SubmitInteractive(ctx context.Context, session Se
 	providerAgentSessionID := a.claudeSDKAdapterSessionID(adapterSession, session.AgentSessionID)
 
 	optionID := strings.TrimSpace(input.OptionID)
+	if optionID == "" && input.Payload != nil {
+		optionID = strings.TrimSpace(asString(input.Payload["optionId"]))
+	}
 	if pending.callType == "approval" {
 		optionID = interactiveApprovalOptionID(input)
 		if optionID == "" {
